feat(api_v2): split program endpoints into spp_api_v2_programs (#1081, #1082)
Extract the Program and ProgramMembership REST API out of spp_api_v2 into a
new spp_api_v2_programs companion (auto-installs when spp_api_v2 + spp_programs
are both present), so spp_api_v2 no longer DIRECTLY depends on spp_programs.

Moved into the companion: program / program_membership routers, services,
schemas, filter configs, api.path records, the program scope-enforcement test,
and the program filter routers (registered into the shared RESOURCE_SERVICES).
Base spp_api_v2 drops the program imports, the spp_programs dependency, and the
program data records.

#1082: spp_api_v2_data has no program code of its own; it only pulled
spp_programs transitively via spp_api_v2. The remaining transitive path
(spp_api_v2 -> spp_source_tracking -> spp_programs) closes in #1084.
