+++ spp_api_v2/schemas/__init__.py
@@ -9,6 +9,4 @@ from . import membership
 from . import operation_outcome
 from . import patch
 from . import problem_detail
-from . import program
-from . import program_membership
 from . import search_result
