+++ spp_api_v2_programs/README.rst
@@ -1,0 +1,113 @@
+=========================
+OpenSPP API V2 - Programs
+=========================
+
+..
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! source digest: sha256:4471b988191c17dbafeaa6a0f0835db915192c85183aabb7f6a6bb66b8f0a213
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Production%2FStable-green.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Production/Stable
+.. |badge2| image:: https://img.shields.io/badge/license-LGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
+    :alt: License: LGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OpenSPP%2FOpenSPP2-lightgray.png?logo=github
+    :target: https://github.com/OpenSPP/OpenSPP2/tree/19.0/spp_api_v2_programs
+    :alt: OpenSPP/OpenSPP2
+
+|badge1| |badge2| |badge3|
+
+Extends OpenSPP API V2 with REST endpoints for **Programs** and
+**Program Memberships**. This module was split out of ``spp_api_v2`` so
+that the base API module no longer depends on ``spp_programs`` —
+deployments that don't use programs (e.g. a registry-only Social
+Registry) can ship the API without installing the Programs stack.
+
+It auto-installs whenever both ``spp_api_v2`` and ``spp_programs`` are
+present, so program API functionality is unchanged for deployments that
+use programs.
+
+Key Capabilities
+~~~~~~~~~~~~~~~~
+
+- **Program endpoints**: read and search programs
+  (``/api/v2/spp/Program``)
+- **Program Membership endpoints**: read, search, create, and update
+  beneficiary enrollments (``/api/v2/spp/ProgramMembership``)
+- **Advanced filtering**: ``/_filters`` and ``/_search`` endpoints for
+  both resources, registered into the shared API V2 filter framework
+- **OAuth scopes**: the ``program`` and ``program_membership``
+  client-scope resources (defined in ``spp_api_v2``) gate access
+
+UI Location
+~~~~~~~~~~~
+
+No standalone menu. Endpoints are available under the API V2 app:
+
+- ``/api/v2/spp/Program``, ``/api/v2/spp/Program/{identifier}``
+- ``/api/v2/spp/ProgramMembership``,
+  ``/api/v2/spp/ProgramMembership/{identifier}``
+
+Dependencies
+~~~~~~~~~~~~
+
+``spp_api_v2``, ``spp_programs``
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Changelog
+=========
+
+19.0.1.0.0
+~~~~~~~~~~
+
+- Split program and program-membership REST endpoints out of
+  ``spp_api_v2`` into this companion module (#1081), so ``spp_api_v2``
+  no longer depends on ``spp_programs``.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OpenSPP/OpenSPP2/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us to smash it by providing a detailed and welcomed
+`feedback <https://github.com/OpenSPP/OpenSPP2/issues/new?body=module:%20spp_api_v2_programs%0Aversion:%2019.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+-------
+
+* OpenSPP.org
+
+Maintainers
+-----------
+
+.. |maintainer-jeremi| image:: https://github.com/jeremi.png?size=40px
+    :target: https://github.com/jeremi
+    :alt: jeremi
+.. |maintainer-gonzalesedwin1123| image:: https://github.com/gonzalesedwin1123.png?size=40px
+    :target: https://github.com/gonzalesedwin1123
+    :alt: gonzalesedwin1123
+.. |maintainer-emjay0921| image:: https://github.com/emjay0921.png?size=40px
+    :target: https://github.com/emjay0921
+    :alt: emjay0921
+
+Current maintainers:
+
+|maintainer-jeremi| |maintainer-gonzalesedwin1123| |maintainer-emjay0921| 
+
+This module is part of the `OpenSPP/OpenSPP2 <https://github.com/OpenSPP/OpenSPP2/tree/19.0/spp_api_v2_programs>`_ project on GitHub.
+
+You are welcome to contribute.
