+++ spp_api_v2_programs/data/api_path_data.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <!-- API Path Configurations for Programs (split from spp_api_v2, OP#1081) -->
+
+    <!-- Program Resource Path -->
+    <record id="api_path_program" model="spp.api.path">
+        <field name="sequence">30</field>
+        <field name="name">Program</field>
+        <field name="model_id" ref="spp_programs.model_spp_program" />
+        <field name="description">Social protection programs</field>
+        <field name="allow_custom_filters">False</field>
+        <field name="max_filter_complexity">10</field>
+    </record>
+
+    <!-- ProgramMembership Resource Path -->
+    <record id="api_path_program_membership" model="spp.api.path">
+        <field name="sequence">40</field>
+        <field name="name">ProgramMembership</field>
+        <field name="model_id" ref="spp_programs.model_spp_program_membership" />
+        <field name="description">Beneficiary enrollments in programs</field>
+        <field name="allow_custom_filters">False</field>
+        <field name="max_filter_complexity">10</field>
+    </record>
+</odoo>
