+++ spp_api_v2_programs/models/__init__.py
@@ -1,0 +1,3 @@
+# Part of OpenSPP. See LICENSE file for full copyright and licensing details.
+
+from . import fastapi_endpoint_programs
