+++ spp_api_v2_programs/readme/HISTORY.md
@@ -1,0 +1,3 @@
+### 19.0.1.0.0
+
+- Split program and program-membership REST endpoints out of `spp_api_v2` into this companion module (#1081), so `spp_api_v2` no longer depends on `spp_programs`.
