+++ spp_api_v2_programs/routers/__init__.py
@@ -1,0 +1,5 @@
+# Part of OpenSPP. See LICENSE file for full copyright and licensing details.
+
+from . import program
+from . import program_filters
+from . import program_membership
