+++ spp_api_v2_programs/routers/program_membership.py
@@ -20,12 +20,12 @@ from fastapi import (
     status,
 )
 
-from ..middleware.auth import get_authenticated_client
+from odoo.addons.spp_api_v2.middleware.auth import get_authenticated_client
 from ..schemas.program_membership import ProgramMembership
-from ..schemas.search_result import SearchResult, create_search_result
-from ..services.consent_service import ConsentService
+from odoo.addons.spp_api_v2.schemas.search_result import SearchResult, create_search_result
+from odoo.addons.spp_api_v2.services.consent_service import ConsentService
 from ..services.program_membership_service import ProgramMembershipService
-from ..utils.pagination import fetch_with_consent
+from odoo.addons.spp_api_v2.utils.pagination import fetch_with_consent
 
 _logger = logging.getLogger(__name__)
