+++ spp_api_v2_programs/schemas/__init__.py
@@ -1,0 +1,4 @@
+# Part of OpenSPP. See LICENSE file for full copyright and licensing details.
+
+from . import program
+from . import program_membership
