+++ spp_api_v2_programs/schemas/program.py
@@ -5,7 +5,7 @@ from typing import Literal
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from .base import CodeableConcept, Identifier, Period, ResourceMeta
+from odoo.addons.spp_api_v2.schemas.base import CodeableConcept, Identifier, Period, ResourceMeta
 
 
 class Program(BaseModel):
