+++ spp_api_v2_programs/tests/__init__.py
@@ -1,0 +1,7 @@
+# Part of OpenSPP. See LICENSE file for full copyright and licensing details.
+
+from . import test_program_api
+from . import test_program_membership_api
+from . import test_program_service
+from . import test_program_membership_service
+from . import test_scope_enforcement_program
