+++ spp_api_v2_programs/tests/test_program_membership_api.py
@@ -4,7 +4,7 @@
 import json
 from datetime import date
 
-from .common import ApiV2HttpTestCase
+from odoo.addons.spp_api_v2.tests.common import ApiV2HttpTestCase
 
 
 class TestProgramMembershipAPIEndpoints(ApiV2HttpTestCase):
